% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootstrap}
\alias{bootstrap}
\title{Bootstraps under SRCM with the 2-stage bootstrap method}
\usage{
bootstrap(surv_data, mle, estimatorfunction, modelfunction, statfunction)
}
\arguments{
\item{surv_data}{the input data, must contain columns 'time', 'event' and 'estimated_survival'}

\item{mle}{the mle of surv_data for the modelfunction}

\item{estimatorfunction}{the function used to estimate the survival for the bootstrap sample,
should be the same which has been called on surv_data before}

\item{modelfunction}{the modelfunction used for resampling under srcm}

\item{statfunction}{the statistics function which is called with surv_data and boot_data}
}
\value{
the value of statfunction(surv_data, boot_data)
}
\description{
Bootstraps under SRCM with the 2-stage bootstrap method
}
