% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_srcm.R
\name{bootstrap_srcm}
\alias{bootstrap_srcm}
\title{Bootstraps statfunction under SRCM with the 2-stage bootstrap method}
\usage{
bootstrap_srcm(statfunction, surv_data, mle, estimatorfunction, modelfunction)
}
\arguments{
\item{statfunction}{the statistics function which is called with surv_data and boot_data}

\item{surv_data}{the input data, must contain columns 'time', 'event' and 'estimated_survival'}

\item{mle}{the mle of surv_data for the modelfunction}

\item{estimatorfunction}{the function used to estimate the survival for the bootstrap sample,
should be the same which has been called on surv_data before}

\item{modelfunction}{the modelfunction used for resampling under srcm}
}
\value{
the value of statfunction(surv_data, boot_data)
}
\description{
Bootstraps statfunction under SRCM with the 2-stage bootstrap method
}
