% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_mle_srcm.R
\name{calculate_mle_srcm}
\alias{calculate_mle_srcm}
\title{Calculates the MLE for the given data and modelfunction under SRCM}
\usage{
calculate_mle_srcm(surv_data, modelfunction)
}
\arguments{
\item{surv_data}{the given survival_data. Needs to have the properties time and event}

\item{modelfunction}{the modelfunction used for the srcm}
}
\value{
the mle of surv_data for the given modelfunction
}
\description{
Calculates the MLE for the given data and modelfunction under SRCM
}
\examples{
surv_data <- dplyr::tibble(time = c(1:5), event = c(1, 0, 1, 0, 1))
calculate_mle_srcm(surv_data, gph)
}
