% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_variance_srcm.R
\name{estimate_variance_srcm}
\alias{estimate_variance_srcm}
\title{Estimates the variance under srcm at all given times of the input sample}
\usage{
estimate_variance_srcm(surv_data, mle, modelfunction, modelfunction_gradient)
}
\arguments{
\item{surv_data}{the input sample, must contain columns "time" and "event"}

\item{mle}{the mle of the given sample and the given model under srcm}

\item{modelfunction}{the modelfunction to be used}

\item{modelfunction_gradient}{the gradient of the modelfunction to be used}
}
\value{
the variance of the given sample at all surv_data$time
}
\description{
Estimates the variance under srcm at all given times of the input sample
}
\examples{
surv_data <- dplyr::tibble(time = c(1:5), event = c(1, 0, 1, 0, 1))
mle <- calculate_mle_srcm(surv_data, gph)
estimate_variance_srcm(surv_data, mle, gph, gph_gradient)
}
