% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimator_dikta16.R
\name{estimator_dikta16}
\alias{estimator_dikta16}
\title{Calculates the semiparametric estimator for survival functions
given by Dikta, Reißel and Harlass (2016)}
\usage{
estimator_dikta16(surv_data, modelfunction, mle)
}
\arguments{
\item{surv_data}{input data, needs to have the properties time and event}

\item{modelfunction}{the modelfunction used for the srcm}

\item{mle}{the mle of surv_data for the modelfunction}
}
\value{
surv_data with additional columns rank and estimated_survival
}
\description{
Calculates the semiparametric estimator for survival functions
given by Dikta, Reißel and Harlass (2016)
}
\examples{
surv_data <- dplyr::tibble(time = c(1:5), event = c(1, 0, 1, 0, 1))
mle <- calculate_mle_srcm(surv_data, gph)
estimator_dikta16(surv_data, gph, mle)
}
