% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimator_dikta96.R
\name{estimator_dikta96}
\alias{estimator_dikta96}
\title{Calculates the semiparametric estimator for the survival function
given by Dikta 96}
\usage{
estimator_dikta96(surv_data, mle, modelfunction)
}
\arguments{
\item{surv_data}{input data, needs to have the properties 'time' and 'event'}

\item{mle}{the mle of surv_data for the modelfunction}

\item{modelfunction}{the modelfunction used for the srcm}
}
\value{
surv_data with additional columns 'rank' and 'estimated_survival'
}
\description{
Calculates the semiparametric estimator for the survival function
given by Dikta 96
}
\examples{
surv_data <- dplyr::tibble(time = c(1:5), event = c(1, 0, 1, 0, 1))
mle <- calculate_mle_srcm(surv_data, gph)
estimator_dikta96(surv_data, mle, gph)
}
