% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_mle_srcm.R
\name{negLL_srcm}
\alias{negLL_srcm}
\title{the negative log-likelihood-function under SRCM}
\usage{
negLL_srcm(par, surv_data, modelfunction)
}
\arguments{
\item{par}{the parameter for modelfunction}

\item{surv_data}{the given survival_data. Needs to have the properties time and event}

\item{modelfunction}{the modelfunction used for the srcm}
}
\value{
the value of the negative log-likelihood-function
}
\description{
the negative log-likelihood-function under SRCM
}
