% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_2stage.R
\name{sample_2stage}
\alias{sample_2stage}
\title{Generates a bootstrap sample with the two-stage bootstrap method}
\usage{
sample_2stage(surv_data, modelfunction, mle)
}
\arguments{
\item{surv_data}{the input sample which is resampled}

\item{modelfunction}{the modelfunction used for resampling}

\item{mle}{the mle of surv_data for the modelfunction}
}
\value{
a bootstrap sample
}
\description{
Generates a bootstrap sample with the two-stage bootstrap method
}
\examples{
surv_data <- dplyr::tibble(time = c(1:5), event = c(1, 0, 1, 0, 1))
mle <- calculate_mle_srcm(surv_data, gph)
sample_2stage(surv_data, gph, mle)
}
