% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCB_enclosed_area_subramanian.R
\name{SCB_enclosed_area_subramanian}
\alias{SCB_enclosed_area_subramanian}
\title{Calculates the enclosed area of the given SCB along the lines of Subramanian
"state the formula here"}
\usage{
SCB_enclosed_area_subramanian(SCB)
}
\arguments{
\item{SCB}{the given SCB, must contain the columns 'time', 'SCB_low' and 'SCB_high'}
}
\value{
the area enclosed by the given SCB
}
\description{
Calculates the enclosed area of the given SCB along the lines of Subramanian
"state the formula here"
}
