% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCB_subramanian_type1.R
\name{SCB_subramanian_type1}
\alias{SCB_subramanian_type1}
\title{Calculates the SCB of type proposed 1 according to Subramanian}
\usage{
SCB_subramanian_type1(
  surv_data,
  estimatorfunction,
  modelfunction,
  alpha,
  n_boot
)
}
\arguments{
\item{surv_data}{the input data, must contain columns 'time' and 'event'}

\item{estimatorfunction}{the plug-in estimator which is used to calculate the SCB}

\item{modelfunction}{the modelfunction used for the srcm}

\item{alpha}{confidence level}

\item{n_boot}{number of bootstrap repititions}
}
\value{
a tibble containing the columns 'time', 'event',
rank', 'estimated_survival', 'SCB_low' and 'SCB_high'
}
\description{
Calculates the SCB of type proposed 1 according to Subramanian
}
\examples{
surv_data <- dplyr::tibble(time = c(1:5), event = c(1, 0, 1, 0, 1))
SCB_subramanian_type1(surv_data, estimator_dikta98, gph, 0.05, 100)
}
