% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCB_subramanian_type2.R
\name{SCB_subramanian_type2}
\alias{SCB_subramanian_type2}
\title{Title}
\usage{
SCB_subramanian_type2(
  surv_data,
  estimatorfunction,
  modelfunction,
  modelfunction_gradient,
  alpha,
  n_boot
)
}
\arguments{
\item{n_boot}{}
}
\description{
Title
}
