% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCB_width_subramanian.R
\name{SCB_width_subramanian}
\alias{SCB_width_subramanian}
\title{Calculates the width of the given SCB along the lines of Subramanian}
\usage{
SCB_width_subramanian(SCB)
}
\arguments{
\item{SCB}{the given SCB, must contain the columns 'estimated_survival', 'SCB_low' and 'SCB_high'}
}
\value{
the width of the given SCB
}
\description{
Calculates the width of the given SCB along the lines of Subramanian
}
