% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sup_Wstar_stat.R
\name{sup_Wstar_stat}
\alias{sup_Wstar_stat}
\title{The statistic which provides the supremum of W*}
\usage{
sup_Wstar_stat(surv_data, boot_data)
}
\arguments{
\item{surv_data}{the original data, must contain columns 'time', 'event' and 'estimated_survival'}

\item{boot_data}{the bootstrapped data, must contain columns 'time', 'event' and 'estimated_survival'}
}
\value{
sup (W* = sqrt(n) * abs(S(t) - S*(t)))
}
\description{
The statistic which provides the supremum of W*
}
