% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twop_cloglog_gradient.R
\name{twop_cloglog_gradient}
\alias{twop_cloglog_gradient}
\title{Returns the gradient values of the twop_cloglog model}
\usage{
twop_cloglog_gradient(t, params)
}
\arguments{
\item{t}{the input times}

\item{params}{the 2 parameters for cloglog model}
}
\value{
value of the twop_cloglog gradient at times t for params
}
\description{
Returns the gradient values of the twop_cloglog model
}
\examples{
twop_cloglog_gradient(0.5, c(1, 1))
}
