% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_SCB.R
\name{draw_SCB}
\alias{draw_SCB}
\title{Title}
\usage{
draw_SCB(
  SCB,
  name = "SCB",
  title = "SCB",
  plt = ggplot2::ggplot(),
  S_true = NULL,
  x_lims = NULL,
  y_lims = NULL,
  my_color = "red"
)
}
\arguments{
\item{y_lims}{}
}
\description{
Title
}
