% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gph.R
\name{gph}
\alias{gph}
\title{Generalized proportional hazard (GPH) model function}
\usage{
gph(t, params)
}
\arguments{
\item{t}{the value of t at which the function should be evaluated}

\item{params}{c(theta1, theta2) of the parameters theta1 and theta2 for gph}
}
\value{
value of the GPH model function with parameters theta1 and theta2
at the given value of t:  \cr \cr
theta1 / (theta1 + t^theta2)
}
\description{
Generalized proportional hazard (GPH) model function
}
\examples{
gph(0.2, params = c(10, 2))
}
