% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCB_covering.R
\name{SCB_covering}
\alias{SCB_covering}
\title{Checks whether the SCB fully covers the true survival function or not}
\usage{
SCB_covering(SCB, true_S)
}
\arguments{
\item{SCB}{the given SCB, must contain the columns 'time', 'SCB_low' and 'SCB_high'}

\item{true_S}{the true survival function}
}
\value{
TRUE if the SCB fully covers the true survival function, FALSE if not
}
\description{
Checks whether the SCB fully covers the true survival function or not
}
