% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCB_subramanian_type1.R
\name{SCB_subramanian_type1}
\alias{SCB_subramanian_type1}
\title{Calculates the SCB of type proposed 1 according to Subramanian}
\usage{
SCB_subramanian_type1(
  surv_data,
  estimatorfunction,
  modelfunction,
  mlefunction,
  alpha,
  n_boot
)
}
\arguments{
\item{n_boot}{}
}
\description{
Calculates the SCB of type proposed 1 according to Subramanian
}
